Getting started adding skeletal meshes to Super Mario Sunshine.
Adding several actors, mostly focused on those in Delfino Plaza.
Adding Support for actors with multiple meshes. This implements a new
ObjectRenderer object that is pretty much an exact copy of the one for
wind waker with some minor tweaks. Will probably want to clean this up
to have less duplicate code.